# openapi.yaml
openapi: 3.0.3
info:
  title: SiteDeck PM API
  description: Project management API for EPC contractors. V1 scope.
  version: 1.0.0
  contact:
    name: Site Deck LLC
servers:
  - url: /api/v1
    description: V1 API
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      summary: List projects
      tags: [Projects]
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      summary: Create project
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, orgId, structureType]
              properties:
                name: { type: string }
                orgId: { type: string }
                structureType: { type: string, enum: [WBS, COST_CODE] }
                startDate: { type: string, format: date-time }
                endDate: { type: string, format: date-time }
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{projectId}:
    get:
      summary: Get project by ID
      tags: [Projects]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    patch:
      summary: Update project
      tags: [Projects]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                status: { type: string }
                startDate: { type: string, format: date-time }
                endDate: { type: string, format: date-time }
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      summary: Delete (cancel) project
      tags: [Projects]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Project cancelled
  /projects/{projectId}/wbs-items:
    post:
      summary: Add WBS item
      tags: [Projects]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name]
              properties:
                code: { type: string }
                name: { type: string }
                parentId: { type: string }
                level: { type: integer }
      responses:
        '201':
          description: WBS item created
...
  /projects/{projectId}/cost/budget-lines:
    get:
      summary: List budget lines
      tags: [Cost]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Budget lines list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BudgetLine'
    post:
      summary: Create budget line
      tags: [Cost]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, budgetAmount]
              properties:
                name: { type: string }
                budgetAmount: { type: number }
                wbsItemId: { type: string }
                costCode: { type: string }
                varianceThreshold: { type: number }
      responses:
        '201':
          description: Budget line created
  /projects/{projectId}/cost/evm:
    get:
      summary: Calculate project EVM
      tags: [Cost]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: EVM results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvmResult'
...
  /projects/{projectId}/procurement/purchase-orders:
    get:
      summary: List purchase orders
      tags: [Procurement]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: PO list
    post:
      summary: Create purchase order
      tags: [Procurement]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [poNumber, vendorName, createdBy, lineItems]
              properties:
                poNumber: { type: string }
                vendorName: { type: string }
                createdBy: { type: string }
                lineItems:
                  type: array
                  items:
                    type: object
                    properties:
                      materialName: { type: string }
                      quantity: { type: number }
                      unit: { type: string }
                      unitPrice: { type: number }
      responses:
        '201':
          description: PO created
  /projects/{projectId}/procurement/subcontracts:
    get:
      summary: List subcontracts
      tags: [Procurement]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Subcontracts list
    post:
      summary: Create subcontract
      tags: [Procurement]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subcontractorName, contractAmount]
              properties:
                subcontractorName: { type: string }
                contractAmount: { type: number }
      responses:
        '201':
          description: Subcontract created
...
  /projects/{projectId}/communications/rfis:
    get:
      summary: List RFIs
      tags: [Communications]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: RFI list
    post:
      summary: Create RFI
      tags: [Communications]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rfiNumber, subject, description, submittedBy]
              properties:
                rfiNumber: { type: string }
                subject: { type: string }
                description: { type: string }
                submittedBy: { type: string }
      responses:
        '201':
          description: RFI created
  /projects/{projectId}/communications/submittals:
    get:
      summary: List submittals
      tags: [Communications]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Submittals list
    post:
      summary: Create submittal
      tags: [Communications]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [submittalNumber, title, submittedBy]
              properties:
                submittalNumber: { type: string }
                title: { type: string }
                description: { type: string }
                submittedBy: { type: string }
      responses:
        '201':
          description: Submittal created
...
  /projects/{projectId}/integration/issues:
    get:
      summary: List issues
      tags: [Integration]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Issues list
    post:
      summary: Create issue
      tags: [Integration]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, source, title, description, createdBy]
              properties:
                type: { type: string }
                source: { type: string }
                title: { type: string }
                description: { type: string }
                createdBy: { type: string }
      responses:
        '201':
          description: Issue created
  /projects/{projectId}/resource/equipment:
    get:
      summary: List equipment
      tags: [Resource]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Equipment list
  /projects/{projectId}/resource/idle-equipment:
    get:
      summary: Idle equipment on critical path
      tags: [Resource]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Idle equipment items
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        orgId: { type: string }
        status: { type: string }
        structureType: { type: string }
        structureLocked: { type: boolean }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    BudgetLine:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        budgetAmount: { type: number }
        committedAmount: { type: number }
        incurredAmount: { type: number }
        percentComplete: { type: number }
        varianceFlag: { type: string }
    EvmResult:
      type: object
      properties:
        projectId: { type: string }
        totalBudget: { type: number }
        totalBcwp: { type: number }
        totalAcwp: { type: number }
        evm:
          type: object
          properties:
            bcws: { type: number }
            bcwp: { type: number }
            acwp: { type: number }
            sv: { type: number }
            cv: { type: number }
            spi: { type: number }
            cpi: { type: number }
            eac: { type: number }
            vac: { type: number }
            tcpi: { type: number }
        lineResults:
          type: array
          items:
            type: object
...